+++ App.js
@@ -1,18 +1,24 @@
 import { createSwitchNavigator, createAppContainer } from 'react-navigation'
 // import the different screens
-import Loading from './screens/AuthLoading'
+import AuthLoading from './screens/AuthLoading'
 import SignUp from './screens/SignUp'
 import Login from './screens/Login'
 import Dashboard from './screens/Dashboard'
+import LaunchDetail from "./screens/LaunchDetail"
+
+// Unfortunately stack navigator with RNv0.60 is broken and cumbersome to fix so I fake implement it my own way.
+// const DashboardStack = createStackNavigator({ Dashboard, LaunchDetail })
+
 // create our app's navigation stack
 export default createAppContainer(createSwitchNavigator(
   {
-    Loading,
+    AuthLoading,
     SignUp,
     Login,
-    Dashboard
+    Dashboard,
+    LaunchDetail
   },
   {
-    initialRouteName: 'Loading'
+    initialRouteName: 'AuthLoading'
   }
 ))