+++ constants.js
@@ -1,6 +1,7 @@
 module.exports = {
   API_LAUNCHES: 'https://launchlibrary.net/1.4/launch',
   API_ROCKETS: 'https://launchlibrary.net/1.4/rocket',
+  API_LAUNCH_STATUS: 'https://launchlibrary.net/1.4/launchstatus',
   URL_DEFAULT_ROCKET_IMG: 'https://image.shutterstock.com/z/stock-vector-rocket-globe-cloud-sky-paper-art-style-with-pastel-color-tones-vector-illustration-765547279.jpg',
   ERROR_MSG_EMPTY_FIELDS: 'Please fill all the fields',
 }